// src/app/model/common/invite.ts
export class Invite {
    invitedByUserId;
    invitedTo: string;
    email: string;
    phone: string;
    inviteType: string;
    businessType: string;
    shareCatelog: boolean;
    catelog: string;
    invitedByBusinessAccountId: string;
    redirectReferenceId: string;
    redirectType: string;
}